services/browserbot/pyproject.toml
[project]
name = "browserbot-service"
version = "0.1.0"
description = "Browser-use FastAPI bridge for Amazon SiteStripe automation"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.9.5",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.0",
    "pydantic>=2.11.5",
    "python-dotenv==1.0.1",
    "pyotp==2.9.0",
]

[tool.uv]
package = true
